+++ lib/linalg.dart
@@ -1,3 +1,8 @@
+export 'package:ml_linalg/axis.dart';
+export 'package:ml_linalg/distance.dart';
+export 'package:ml_linalg/dtype.dart';
 export 'package:ml_linalg/matrix.dart';
+export 'package:ml_linalg/matrix_norm.dart';
 export 'package:ml_linalg/norm.dart';
+export 'package:ml_linalg/sort_direction.dart';
 export 'package:ml_linalg/vector.dart';