+++ lib/src/common/helper/get_length_of_first_or_zero.dart
@@ -1,0 +1,2 @@
+int getLengthOfFirstOrZero<T extends Iterable<double>>(List<T> collection) =>
+    collection.isNotEmpty ? collection.first.length : 0;