+++ test/float32_matrix_test.dart
@@ -1,6 +1,8 @@
 import 'dart:typed_data';
 
+import 'package:ml_linalg/axis.dart';
 import 'package:ml_linalg/matrix_norm.dart';
+import 'package:ml_linalg/sort_direction.dart';
 import 'package:ml_linalg/src/matrix/float32/float32_matrix.dart';
 import 'package:ml_linalg/src/vector/float32/float32_vector.dart';
 import 'package:ml_linalg/vector.dart';
@@ -23,6 +25,15 @@ void main() {
       expect(actual.columnsNum, 5);
     });
 
+    test('should create an instance based on an empty list (`fromList` '
+        'constructor)', () {
+      final actual = Float32Matrix.fromList([]);
+      final expected = <double>[];
+      expect(actual, equals(expected));
+      expect(actual.rowsNum, 0);
+      expect(actual.columnsNum, 0);
+    });
+
     test('should create an instance with predefined vectors as matrix '
         'rows', () {
       final actual = Float32Matrix.rows([
@@ -36,6 +47,15 @@ void main() {
       expect(actual, equals(expected));
       expect(actual.rowsNum, 2);
       expect(actual.columnsNum, 5);
+    });
+
+    test('should create an instance based on an empty list (`rows` '
+        'constructor)', () {
+      final actual = Float32Matrix.rows([]);
+      final expected = <double>[];
+      expect(actual, equals(expected));
+      expect(actual.rowsNum, 0);
+      expect(actual.columnsNum, 0);
     });
 
     test('should create an instance with predefined vectors as matrix columns',
@@ -56,6 +76,15 @@ void main() {
       expect(actual.columnsNum, 2);
     });
 
+    test('should create an instance based on an empty list (`columns` '
+        'constructor)', () {
+      final actual = Float32Matrix.columns([]);
+      final expected = <double>[];
+      expect(actual, equals(expected));
+      expect(actual.rowsNum, 0);
+      expect(actual.columnsNum, 0);
+    });
+
     test('should create an instance from flattened collection', () {
       final actual =
           Float32Matrix.flattened([1.0, 2.0, 3.0, 4.0, 5.0, 6.0], 2, 3);
@@ -65,6 +94,16 @@ void main() {
       ];
       expect(actual.rowsNum, 2);
       expect(actual.columnsNum, 3);
+      expect(actual, equals(expected));
+    });
+
+    test('should create an instance based on an empty list (`flattened` '
+        'constructor)', () {
+      final actual =
+      Float32Matrix.flattened([], 0, 0);
+      final expected = <double>[];
+      expect(actual.rowsNum, 0);
+      expect(actual.columnsNum, 0);
       expect(actual, equals(expected));
     });
 
@@ -1309,6 +1348,92 @@ void main() {
         [112.0, 10.0, 34.0, 2.0, 10.0],
       ];
       expect(actual, equals(expected));
+    });
+  });
+
+  group('Float32x4Matrix.sort', () {
+    test('should sort the matrix row-wise with asc direction', () {
+      final matrix = Float32Matrix.fromList([
+        [4.0, 8.0, 12.0, 16.0, 34.0],
+        [20.0, 24.0, 28.0, 32.0, 23.0],
+        [36.0, .0, -8.0, -12.0, 12.0],
+        [16.0, 1.0, -18.0, 3.0, 11.0],
+        [112.0, 10.0, 34.0, 2.0, 10.0],
+      ]);
+      final expected = [
+        [16.0, 1.0, -18.0, 3.0, 11.0],
+        [36.0, .0, -8.0, -12.0, 12.0],
+        [4.0, 8.0, 12.0, 16.0, 34.0],
+        [20.0, 24.0, 28.0, 32.0, 23.0],
+        [112.0, 10.0, 34.0, 2.0, 10.0],
+      ];
+      final actual = matrix.sort((vector) => vector[2], Axis.rows,
+          SortDirection.asc);
+      expect(actual, equals(expected));
+      expect(matrix, isNot(equals(expected)));
+    });
+
+    test('should sort the matrix column-wise with asc direction', () {
+      final matrix = Float32Matrix.fromList([
+        [4.0, 8.0, 12.0, 16.0, 34.0],
+        [20.0, 24.0, 28.0, 32.0, 23.0],
+        [36.0, .0, -8.0, -12.0, 12.0],
+        [16.0, 1.0, -18.0, 3.0, 11.0],
+        [112.0, 10.0, 34.0, 2.0, 10.0],
+      ]);
+      final expected = [
+        [16.0, 12.0, 8.0, 34.0, 4.0],
+        [32.0, 28.0, 24.0, 23.0, 20.0],
+        [-12.0, -8.0, 0.0, 12.0, 36.0],
+        [3.0, -18.0, 1.0, 11.0, 16.0],
+        [2.0, 34.0, 10.0, 10.0, 112.0],
+      ];
+      final actual = matrix.sort((vector) => vector[2], Axis.columns,
+          SortDirection.asc);
+      expect(actual, equals(expected));
+      expect(matrix, isNot(equals(expected)));
+    });
+
+    test('should sort the matrix row-wise with desc direction', () {
+      final matrix = Float32Matrix.fromList([
+        [4.0, 8.0, 12.0, 16.0, 34.0],
+        [20.0, 24.0, 28.0, 32.0, 23.0],
+        [36.0, .0, -8.0, -12.0, 12.0],
+        [16.0, 1.0, -18.0, 3.0, 11.0],
+        [112.0, 10.0, 34.0, 2.0, 10.0],
+      ]);
+      final expected = [
+        [112.0, 10.0, 34.0, 2.0, 10.0],
+        [20.0, 24.0, 28.0, 32.0, 23.0],
+        [4.0, 8.0, 12.0, 16.0, 34.0],
+        [36.0, .0, -8.0, -12.0, 12.0],
+        [16.0, 1.0, -18.0, 3.0, 11.0],
+      ];
+      final actual = matrix.sort((vector) => vector[2], Axis.rows,
+          SortDirection.desc);
+      expect(actual, equals(expected));
+      expect(matrix, isNot(equals(expected)));
+    });
+
+    test('should sort the matrix column-wise with desc direction', () {
+      final matrix = Float32Matrix.fromList([
+        [4.0, 8.0, 12.0, 16.0, 34.0],
+        [20.0, 24.0, 28.0, 32.0, 23.0],
+        [36.0, .0, -8.0, -12.0, 12.0],
+        [16.0, 1.0, -18.0, 3.0, 11.0],
+        [112.0, 10.0, 34.0, 2.0, 10.0],
+      ]);
+      final expected = [
+        [4.0, 34.0, 8.0, 12.0, 16.0],
+        [20.0, 23.0, 24.0, 28.0, 32.0],
+        [36.0, 12.0, 0.0, -8.0, -12.0],
+        [16.0, 11.0, 1.0, -18.0, 3.0],
+        [112.0, 10.0, 10.0, 34.0, 2.0],
+      ];
+      final actual = matrix.sort((vector) => vector[2], Axis.columns,
+          SortDirection.desc);
+      expect(actual, equals(expected));
+      expect(matrix, isNot(equals(expected)));
     });
   });