+++ test/float32_vector_test.dart
@@ -5,12 +5,13 @@ import 'package:ml_linalg/linalg.dart';
 import 'package:ml_linalg/src/matrix/float32/float32_matrix.dart';
 import 'package:ml_linalg/src/vector/float32/float32_vector.dart';
 import 'package:test/test.dart';
+import 'package:xrange/zrange.dart';
 
 import 'unit_test_helpers/float_iterable_almost_equal_to.dart';
 
 void main() {
   group('Float32Vector', () {
-    group('`from` constructor', () {
+    group('`fromList` constructor', () {
       test('should create a vector from dynamic-length list, length is '
           'greater than 4', () {
         final vector1 = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0, 6.0]);
@@ -38,6 +39,12 @@ void main() {
         final vector = Float32Vector.fromList(List.filled(1, 2.0));
         expect(vector, equals([2.0]));
         expect(vector.length, 1);
+      });
+
+      test('should create a vector from an empty list', () {
+        final vector = Float32Vector.fromList([]);
+        expect(vector, equals(<double>[]));
+        expect(vector.length, 0);
       });
     });
 
@@ -73,6 +80,40 @@ void main() {
         expect(
             vector, equals([2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0]));
         expect(vector.length, equals(10));
+      });
+    });
+
+    group('`randomFilled` constructor', () {
+      test('should create a vector filled with random values from range'
+          '[-5; -1)', () {
+        final vector = Float32Vector.randomFilled(200, min: -5, max: -1);
+        for (final element in vector) {
+          expect(element, inClosedOpenRange(-5, -1));
+        }
+      });
+
+      test('should create a vector filled with random values from range'
+          '[-5; 10)', () {
+        final vector = Float32Vector.randomFilled(200, min: -5, max: 10);
+        for (final element in vector) {
+          expect(element, inClosedOpenRange(-5, 10));
+        }
+      });
+
+      test('should create a vector filled with random values from range'
+          '[-5; -1) (min is greater than max)', () {
+        final vector = Float32Vector.randomFilled(200, min: -1, max: -5);
+        for (final element in vector) {
+          expect(element, inClosedOpenRange(-5, -1));
+        }
+      });
+
+      test('should create a vector filled with constant value if min equals '
+          'max', () {
+        final vector = Float32Vector.randomFilled(200, min: 2, max: 2);
+        for (final element in vector) {
+          expect(element, equals(2));
+        }
       });
     });
 
@@ -167,6 +208,7 @@ void main() {
         'object is not a vector', () {
       final vector1 = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0, 6.0]);
       final vector2 = [1.0, 2.0, 3.0, 4.0, 5.0, 6.0];
+      // ignore: unrelated_type_equality_checks
       expect(vector1 == vector2, isFalse);
     });
 
@@ -454,7 +496,8 @@ void main() {
       expect(result, isNot(vector));
     });
 
-    test('`query` method', () {
+    test('should create a vector using elements on specific inidces from '
+        'given list', () {
       final vector = Float32Vector.fromList([10.0, 3.0, 4.0, 7.0, 9.0, 12.0]);
       final query = vector.query([1, 1, 0, 3]);
       expect(query, equals([3.0, 3.0, 10.0, 7.0]));
@@ -561,12 +604,166 @@ void main() {
       expect(() => vector[1], throwsRangeError);
       expect(() => vector[100], throwsRangeError);
     });
+  });
 
-    test('should cut out a subvector', () {
+  group('Float32x4Vector.subvector', () {
+    test('should cut out a subvector (`end` exclusive)', () {
       final vector = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0]);
-      final actual = vector.subvector(1, 5);
-      final expected = [2.0, 3.0, 4.0, 5.0];
+      final actual = vector.subvector(1, 4);
+      final expected = [2.0, 3.0, 4.0];
       expect(actual, expected);
+    });
+
+    test('should cut out a subvector of length 1 if `start` is equal to the '
+        'last index of the vector', () {
+      final vector = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0]);
+      final actual = vector.subvector(4, 5);
+      final expected = [5.0];
+      expect(actual, expected);
+    });
+
+    test('should cut out rest of the vector if `end` is not specified', () {
+      final vector = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0, 7.0]);
+      final actual = vector.subvector(1);
+      final expected = [2.0, 3.0, 4.0, 5.0, 7.0];
+      expect(actual, expected);
+    });
+
+    test('should cut out rest of the vector if `end` is specified and greater'
+        'that the vector length', () {
+      final vector = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0, 7.0]);
+      final actual = vector.subvector(1, 20);
+      final expected = [2.0, 3.0, 4.0, 5.0, 7.0];
+      expect(actual, expected);
+    });
+
+    test('should throw a range error if `start` is negative', () {
+      final vector = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0, 7.0]);
+      final actual = () => vector.subvector(-1, 20);
+      expect(actual, throwsRangeError);
+    });
+
+    test('should throw a range error if `start` is greater than `end`', () {
+      final vector = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0, 7.0]);
+      final actual = () => vector.subvector(3, 2);
+      expect(actual, throwsRangeError);
+    });
+
+    test('should throw a range error if `start` is equal to the `end`', () {
+      final vector = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0]);
+      final actual = () => vector.subvector(4, 4);
+      expect(actual, throwsRangeError);
+    });
+  });
+
+  group('Float32x4Vector.subvectorByRange', () {
+    test('should return a proper vector using closed range', () {
+      final vector = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0]);
+      final actual = vector.subvectorByRange(ZRange.closed(1, 3));
+      expect(actual, equals([2.0, 3.0, 4.0]));
+    });
+
+    test('should return a proper vector using open range', () {
+      final vector = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0]);
+      final actual = vector.subvectorByRange(ZRange.open(1, 3));
+      expect(actual, equals([3.0]));
+    });
+
+    test('should return a proper vector using open-closed range', () {
+      final vector = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0]);
+      final actual = vector.subvectorByRange(ZRange.openClosed(1, 3));
+      expect(actual, equals([3.0, 4.0]));
+    });
+
+    test('should return a proper vector using closed-open range', () {
+      final vector = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0]);
+      final actual = vector.subvectorByRange(ZRange.closedOpen(1, 3));
+      expect(actual, equals([2.0, 3.0]));
+    });
+
+    test('should return a proper vector using upper boundless range', () {
+      final vector = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0]);
+      final actual = vector.subvectorByRange(ZRange.atLeast(2));
+      expect(actual, equals([3.0, 4.0, 5.0]));
+    });
+
+    test('should return a proper vector using lower boundless range', () {
+      final vector = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0]);
+      final actual = vector.subvectorByRange(ZRange.atMost(2));
+      expect(actual, equals([1.0, 2.0, 3.0]));
+    });
+
+    test('should return a copy of the vector using absolutely boundless '
+        'range', () {
+      final vector = Float32Vector.fromList([1.0, 2.0, 3.0, 4.0, 5.0]);
+      final actual = vector.subvectorByRange(ZRange.all());
+      expect(actual, equals([1.0, 2.0, 3.0, 4.0, 5.0]));
+    });
+  });
+
+  group('Float32x4Vector.hashCode', () {
+    test('should return the same hashcode for equal vectors, case 1', () {
+      final hash1 = Float32Vector.fromList([0, 0, 0, 0, 1]).hashCode;
+      final hash2 = Float32Vector.fromList([0, 0, 0, 0, 1]).hashCode;
+      expect(hash1, equals(hash2));
+    });
+
+    test('should return the same hashcode for equal vectors, case 2', () {
+      final hash1 = Float32Vector.fromList([-10, double.infinity, 345, 20, 1])
+          .hashCode;
+      final hash2 = Float32Vector.fromList([-10, double.infinity, 345, 20, 1])
+          .hashCode;
+      expect(hash1, equals(hash2));
+    });
+
+    test('should return the same hashcode for equal vectors, case 3', () {
+      final hash1 = Float32Vector.fromList([0, 0, 0, 0, 0]).hashCode;
+      final hash2 = Float32Vector.fromList([0, 0, 0, 0, 0]).hashCode;
+      expect(hash1, equals(hash2));
+    });
+
+    test('should return the same hashcode for equal vectors, case 4', () {
+      final hash1 = Float32Vector.fromList([]).hashCode;
+      final hash2 = Float32Vector.fromList([]).hashCode;
+      expect(hash1, equals(hash2));
+    });
+
+    test('should return the same hashcode for equal vectors, case 5', () {
+      final hash1 = Float32Vector.fromList([100]).hashCode;
+      final hash2 = Float32Vector.fromList([100]).hashCode;
+      expect(hash1, equals(hash2));
+    });
+
+    test('should return a different hashcode for unequal vectors, case 1', () {
+      final hash1 = Float32Vector.fromList([0, 0, 0, 1, 0]).hashCode;
+      final hash2 = Float32Vector.fromList([0, 0, 0, 0, 1]).hashCode;
+      expect(hash1, isNot(equals(hash2)));
+    });
+
+    test('should return a different hashcode for unequal vectors, case 2', () {
+      final hash1 = Float32Vector.fromList([0, 0, 0, 10, 0]).hashCode;
+      final hash2 = Float32Vector.fromList([0, 0, 0, 1, 0, 0]).hashCode;
+      expect(hash1, isNot(equals(hash2)));
+    });
+
+    test('should return a different hashcode for unequal vectors, case 3', () {
+      final hash1 = Float32Vector.fromList([-32, 12, 0, 10, 0]).hashCode;
+      final hash2 = Float32Vector.fromList([32, 12, 0, 10, 0]).hashCode;
+      expect(hash1, isNot(equals(hash2)));
+    });
+
+    test('should return a different hashcode for unequal vectors, case 4', () {
+      final hash1 = Float32Vector.fromList([32, 5, 46, 78, 9]).hashCode;
+      final hash2 = Float32Vector.fromList([3, 25, 46, 7, 89]).hashCode;
+      expect(hash1, isNot(equals(hash2)));
+    });
+
+    test('should return a different hashcode for unequal vectors, case 5', () {
+      final hash1 = Float32Vector.fromList([32.04999923706055, 0.5, 2.0, 11.5])
+          .hashCode;
+      final hash2 = Float32Vector.fromList([32.0, 49999237060550.5, 2.0, 11.5])
+          .hashCode;
+      expect(hash1, isNot(equals(hash2)));
     });
   });
 }